chore: simplify back-end services
