+++ server/src/main/java/br/com/tasknoteapp/server/controller/UserSessionController.java
@@ -11,6 +11,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -48,5 +49,10 @@ public class UserSessionController {
   public JwtAuthenticationResponse refresh() {
     String token = authService.refreshCurrentUserToken();
     return new JwtAuthenticationResponse(token);
+  }
+
+  @PostMapping("/delete-account")
+  public void deteleAccount() {
+    //
   }
 }
