+++ server/src/main/java/br/com/tasknoteapp/server/service/UserSessionService.java
@@ -1,0 +1,13 @@
+package br.com.tasknoteapp.server.service;
+
+import org.springframework.stereotype.Service;
+import jakarta.transaction.Transactional;
+
+@Service
+public class UserSessionService {
+    
+  @Transactional
+  public void deleteCurrentUserAccount() {
+    //
+  }
+}
